Refactor RepositoryMainInfo.tsx
